+++ src/Payment/resources/database/migrations/2023_03_06_205137_add_subtype_to_payments_table.php
@@ -1,0 +1,24 @@
+<?php
+
+declare(strict_types=1);
+
+use Illuminate\Database\Migrations\Migration;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\Schema;
+
+return new class() extends Migration
+{
+    public function up()
+    {
+        Schema::table('payments', function (Blueprint $table) {
+            $table->string('subtype')->nullable()->after('payment_method_id');
+        });
+    }
+
+    public function down()
+    {
+        Schema::table('payments', function (Blueprint $table) {
+            $table->dropColumn('subtype');
+        });
+    }
+};