+++ src/Payment/Changelog.md
@@ -2,6 +2,11 @@
 
 ## 3.x Series
 
+## Unreleased
+##### 2023-XX-YY
+
+- Added the `subtype` field and the `getSubtype()` method to the `Payment` model
+
 ## 3.5.0
 ##### 2023-02-23